Review corrections

+++ src/uci.cpp
@@ -183,25 +183,25 @@ namespace {
          << "\nNodes/second    : " << 1000 * nodes / elapsed << endl;
   }
 
-  // The win rate model return the probability (per mille) of winning given
-  //  a game-ply and eval. The model fits rather accurately the LTC fishtest statistics.
+  // The win rate model returns the probability (per mille) of winning given an eval
+  // and a game-ply. The model fits rather accurately the LTC fishtest statistics.
   int win_rate_model(Value v, int ply) {
 
-     // The model captures only up to 240 plies, so limit input (and rescale).
-     float m = std::min(240, ply) / 64.0;
+     // The model captures only up to 240 plies, so limit input (and rescale)
+     double m = std::min(240, ply) / 64.0;
 
      // Coefficients of a 3rd order polynomial fit based on fishtest data
      // for two parameters needed to transform eval to the argument of a
      // logistic function.
-     float as[] = {-8.24404295, 64.23892342, -95.73056462, 153.86478679};
-     float bs[] = {-3.37154371, 28.44489198, -56.67657741,  72.05858751};
-     float a = (((as[0] * m + as[1]) * m + as[2]) * m) + as[3];
-     float b = (((bs[0] * m + bs[1]) * m + bs[2]) * m) + bs[3];
+     double as[] = {-8.24404295, 64.23892342, -95.73056462, 153.86478679};
+     double bs[] = {-3.37154371, 28.44489198, -56.67657741,  72.05858751};
+     double a = (((as[0] * m + as[1]) * m + as[2]) * m) + as[3];
+     double b = (((bs[0] * m + bs[1]) * m + bs[2]) * m) + bs[3];
 
-     // transform eval to cp with limited range
-     float x = Utility::clamp(float(100 * v) / PawnValueEg, -1000.0f, 1000.0f);
+     // Transform eval to centipawns with limited range
+     double x = Utility::clamp(double(100 * v) / PawnValueEg, -1000.0, 1000.0);
 
-     // return win rate in per mille (rounded to nearest)
+     // Return win rate in per mille (rounded to nearest)
      return int(0.5 + 1000 / (1 + std::exp((a - x) / b)));
   }
